Added ability to specify mods in modstat command

+++ factorio.py
@@ -6,10 +6,14 @@ class FactorioCog(commands.Cog, name='Factorio'):
         self.bot = bot
 
     @commands.command(aliases=['modstat'], description='Info about mods', brief='Info about mods', help='Prints a bunch of commands for uBot to display info about mods')
-    async def mods_statistics(self, ctx):
-        for modname in MOD_LIST:
-            await ctx.send(content='>>{0}<<'.format(modname), delete_after=1)
-        await ctx.message.delete()
+    async def mods_statistics(self, ctx, *mod_names):
+        if mod_names:
+            for modname in mod_names:
+                await ctx.send('>>'+modname+'<<', delete_after=1)
+        else:
+            for modname in MOD_LIST:
+                await ctx.send(content='>>{0}<<'.format(modname), delete_after=1)
+            await ctx.message.delete()
 
 def setup(bot):
     bot.add_cog(FactorioCog(bot))
